refactor(datasource graphql hasura): split aggregation and polymorphism detection out
The collection carried the whole aggregation pipeline (validation, the parent-table
detour Hasura forces on grouped aggregates, value coercion) and the introspector
carried the polymorphism detection. Both now live in classes of their own,
Query::Aggregator and Introspection::PolymorphismDetector, leaving the collection
to its CRUD surface and the introspector to reading the schema.

Configuration takes its options as a keyword hash validated against the known
list, so an unknown option is reported by name.
